requirements.txt
streamlit==1.49.1
openai>=1.40.0
numpy==1.26.4
faiss-cpu==1.7.4
pyarrow==11.0.0
